Update build wallet.md
This is a batch paper wallet for Polkadot (and btc, eth and eos).  It can generate 1 to 1000 wallets and export as pdf. It using combination of drawing on canvas (mouse movement) and Polkadot-js api random hex function to generate raw seed. This method will increase the entropy and make the private keys generated more random and secured.