Update maintain-polkadot-parameters.md
As this is my first update to the Wiki, I am making sure I get the convention right

Do we link to referendums and include dates when the changes happen? 

Also, is this the correct location where we update the details or should we update this in FAQ page? Thanks!