Update maintain-polkadot-parameters.md (#2289)
* Update maintain-polkadot-parameters.md

As this is my first update to the Wiki, I am making sure I get the convention right

Do we link to referendums and include dates when the changes happen? 

Also, is this the correct location where we update the details or should we update this in FAQ page? Thanks!

* Instructions pointing to chain state

* point to polkadot.js for minimum bonding parameter

From the suggestions provided in #2282 I removed the hyperlinks that point to the scripts that my be outdated and instead, pointed it to chain state available through UI

* Update learn-nominator.md

min staking parameter may not change every era