Re-write and renaming of "How to identify legitimate projects"
A re-write and renaming of the article after discussions with the Ecosystem development team. The new article focus more on how to look for and verify information given by a project (hence the new title).

Also, the ID of the article has changed. Is it possible to do a redirect for old links that point to the old URL?