Add AirGap Wallet
AirGap is a secure mobile wallet and has been supported by the Web3Foundation to integrate DOT. (Not through the treasury). We are actively working on both the apps in general and also the DOT integration.

The AirGap Vault application is an offline application similar to the Parity Signer.
The AirGap Wallet application is a companion app to the Vault, which is used to prepare and broadcast transactions, but does not hold any private keys, similar to the polkadot-js apps.

If you need any other information to merge this PR, please let me know.