Update documentation for the Future error
Just adding some additional information for the "Future" error mentioned on the list of common errors. Unlike the rest of the errors on this list, the Future error does not result in the incoming TX getting discarded immediately. Instead, it is moved to the "futures queue" where it may still get executed if additional transactions are executed that use the missing nonces.