Suggestion on toggle to unbond all funds (#3659)
Sorry to be the Wiki pest these days...

We were troubleshooting an issue with a user on discord and through this discussion I'm making some of these updates.  It appears the user manually typed in the amount to unbond leaving some trace KSM in the pool.  The pool still recognized him as a member and didn't allow him to join another.  

While I didn't make the adjustment, could you please confirm under limitations that a member can have 16 partial unbonds, I think this might be 8 but I am not confident.

```
//! In order to leave, a member must take two steps.
//!
//! First, they must call [`Call::unbond`]. The unbond extrinsic will start the unbonding process by
//! unbonding all or a portion of the members funds.
//!
//! > A member can have up to [`Config::MaxUnbonding`] distinct active unbonding requests.
//!
//! Second, once [`sp_staking::StakingInterface::bonding_duration`] eras have passed, the member can
//! call [`Call::withdraw_unbonded`] to withdraw any funds that are free.

pub static MaxUnbonding: u32 = 8
```
[Link to code reference](https://github.com/paritytech/substrate/blob/4a19d408095ebcb47840ad1f2f086f53a0020cb7/frame/nomination-pools/src/mock.rs#L31)