Remove "undetected" when discussing the Monero inflation bug
The presence of torsion is detectable, and the Monero blockchain has 
been confirmed to have not been exploited. This statement, as-is, casted 
doubt onto the current legitimacy of its monetary status with a faulty 
premise.

I do not speak Chinese and used Google Translate to remove characters in 
a similar location until Google Translate no longer said "undetected". I 
cannot confirm I did it properly, nor can I confirm the Chinese is 
correct in general. Google Translate claims it's all present tense, 
despite this being a historical footnote (and a valuable one at that).