Remove the misconception of 100 parachian limit (#4549)
This doc on the Wiki is cited as a reference stating that the number of parachains that can be supported is limited to 100.

The current upper number of Parachains is just an educated guess, something to strive for. Depending on how the system performs in real life, the number could maybe be higher. However, with the introduction of parathreads this number will clearly be higher. This upper number is the number of concurrently advancing parachains/parathreads on the relay chain. As the idea behind parathreads is that certain chains don’t need a fixed block time of 6s and are having varying depending on activity on the parathread, the relay chain can run much more parathreads.

There are also ideas around scaling Polkadot. It isn’t the idea (and I think it never wasn’t) to stop at 100 Parachains. However, at some point we will reach a limit and need to either create “relay chain shards” or maybe create secondary parachains. All of these ideas are theoretical so far, but we are already thinking of what comes after the (first) 100.

https://forum.polkadot.network/t/polkadot-the-centralized-decentralized-ecosystem-a-cartel/1412